Total Reset works. Completed branch.
Moved the totalResetCounter to initialize outside the x for loop. Realized it was resetting each time it moved back an X value if the line was broken.  The branch now works. Though I would rather it not have to reset on breaks, but an algorithm change in the future seems like a fair exchange for something that works practically now.